Updated index.html and scripts.js to begin creating the grid
This commit contains updates to create the grid container in the index.html as well as the beginnings of creating the grid within the scripts.js file.
I have not created the function in full to generate the grid as I'm still working out how to do that.

Changes to be committed:
	modified:   index.html
	new file:   reset.css
	modified:   scripts.js