Created function for creating a grid
This commit includes a function to create a grid based on gridSize which will eventually become a user input. Currently, it's just a variable that's set in code.
This commit also includes base CSS to allow me to visualize the grid without styling at this time.

Changes to be committed:
	modified:   index.html
	modified:   scripts.js
	modified:   styles.css