Updated CSS and scripts.js to change colors when hovering
Updated scripts.js to include a new function of removeChildNodes to reset when grid creation is called.

Added a class of hover-bg to the css and added an eventListener to scripts.js for mouseover events. When a mouseover event is fired, the background color changes to a hardcoded color. This will
be updated in a future version to allowe customization.

Changes to be committed:
	modified:   scripts.js
	modified:   styles.css