Added dynamic grid sizing and better border implementation
With this change, I have updated the outside border to be 1em. This allows the inner border
to be better displayed with all size grids. This change also includes dynamic grid sizing.
I had to update the grid container to a set size and the grid cells to be box sized, border boxing.

Changes to be committed:
	modified:   index.html
	modified:   scripts.js
	modified:   styles.css